Add Content Block Postal Address type
The Content Modelling Team are working on a spike to add the ability to publishing reusable pieces of content (“Content Blocks”) via Whitehall. This adds a simple initial schema to help us work with a type of postal address.

There is an existing `post_addresses` object in the `contact` schema which is shaped slightly differently - I thought it better to go with the shape that GDS currently suggestions in their docs https://design-system.service.gov.uk/patterns/addresses/ to make it easier for our creation forms to comply with GDS standards.

We could add/change these fields in the future.