Add sidekiq ui
This adds the sidekiq UI to publishing API. This is deliberately done in a
really straightforward way. As it stands Whitehall is the only real precident
for this work and work has been done there to hide sidekiq behind a
permission in signon. This isn't required by publishing API though because it
doesn't have a UI or any ingress channels, nor will any have to be added.
Instead we'll have to use port forwarding within K8s to access the interface.
This acts as the authentication layer because only those with AWS access to
our clusters will be able to do that.

There was some discussion about whether we'd be able to get access to our
sidekiq queues through logging or grafana, but this was seen as the lighter
weight option because querying the logs is technically fairly tricky.

I've had to add the ability to have sessions as well or the sidekiq UI just
won't load.