A new query to get content that uses a reusable content block
Naming of this class “GetContentByEditionLink “ isn’t straight forward. “content” in this context refers to both the traditional edition as well as our new reusable pieces of content. We use one piece of __content__ to find other pieces of __content__.

It’s also complicated by the use of links to connect these two concepts, and the array of existing query classes with “link”, “content”, “linkables”, “linked”, “link_set”. Trying to communicate how this class is different without compounding the complexity of links is difficult and may require more thought.

How do we feel about introducing “EditionLink” that is spoken about, into class terminology like this?