New endpoint to return content that’s embedding reusable content
This routing feels restful. We start with the content_id of a resuable document and wish to find all other pieces of __content__ that are linking to it.

```
/content/:content_id/linked-items
```

I’m not set on “linked-items”. Link is the core association by which we achieve this however “link” is heavily used elsewhere. As the reusable content team, do we want to introduce a term like “embedded-content”?

```
/content/:content_id/embedded-content-items
```