Find linked content now returns the publishing organisation
This is a draft idea to prove it can be done. It includes inefficiencies since we have to go and query for every organisation document and it’s latest edition. We’d like to weave this into the initial query.

In dev we found it was possible to end up in the situation where a document didn’t have a publishing document, so we have to account for nil. This may have been because the whitehall organisation that we’re set up with doesn’t exist in the publishing API so the link isn’t created.