Add a rake task to push database metrics to prometheus
I often find myself wanting to know "how many published editions are
there for this document type?", or "how many published editions do we
have for each publishing app?", or things of that nature.

Currently these questions can be answered at a point in time by querying
a database, but there's no easy way to see how the values change over
time.

It's also not easy to compare numbers from different databases - for
example, is the number of editions in publishing-api's database with
publishing_app=whitehall the same of the number of editions in the same
state in whitehall?

The idea here is to have each publishing app and publishing-api report a
metric of "editions_in_database_total". Each publishing app should set
the "database" label to be the app name, and set the other labels as
consistently as possible with publishing-api.

We'll run this rake task as a cron job, and it will push metrics to
prometheus via the pushgateway. The rake task also renders the metrics
as text, to make debugging and testing easier.

The database query in publishing-api is fairly expensive (it has to do a
large heap scan of editions, which takes around 30 seconds), so we
shouldn't run the cron job more than around once an hour.

[Explain analyse output for the query](https://www.pgexplain.dev/plan/466c33ba-3358-4b56-aaaa-8f18a396ec3d)

Currently the output of the task looks like this:

    Found 2103 combinations of labels
    # TYPE editions_in_database_total gauge
    # HELP editions_in_database_total Count of editions in various databases labeled by state, document_type etc.
    editions_in_database_total{database="publishing-api",state="draft",content_store="draft",document_type="aaib_report",publishing_app="specialist-publisher",locale="en"} 34.0
    editions_in_database_total{database="publishing-api",state="draft",content_store="draft",document_type="about",publishing_app="whitehall",locale="cy"} 1.0
    ...

2103 combinations of labels sounds like a lot, but Prometheus is fine
with much larger cardinalities than this, at least according to
[this answer](https://stackoverflow.com/a/69167162)