Get Embedded content query
Given a target_content_id which is the edition of the content block/embeddable content, it will find and return any other live editions where it has been embedded and a link set up.

For The content Object store within Whitehall we only need a few fields compared to the existing `content/:id` endpoint: title, base_path, document_type, target_content_id and a couple of fields about the associated organisation.

With this data we’ll present a table that allows resuable content publishers to see where their content is currently being included.

We have previously tried to use the content/:id endpoint. Although it includes a lot more data than we need it did work locally and in particular loaded the organisation through link expansion. Importantly on Integration this timed out.

We tracked it down to a subquery used[1] when the `link_set_links` field is passed through. We did some local benchmarking and found the query was slowing down notably after a few hundred thousand records:

```
Benchmarking
2024-08-15 17:24:20 Document: 178924
2024-08-15 17:24:20 Edition: 178925
2024-08-15 17:24:20 Link: 178855
2024-08-15 17:24:20   0.003363   0.012408   0.015771 (  0.334085)
1.92s in the browser

2024-08-15 17:29:01 Document: 228173
2024-08-15 17:29:01 Edition: 228174
2024-08-15 17:29:01 Link: 228106
2024-08-15 17:29:01   0.019349   0.013283   0.032632 (  0.499379)
3.8s in the browser
```

We stopped seeding at this point as we could see that seeding more would eventually result in a timeout.

On integration the comparable data counts at the time were:

```
39_953_449 links
14_170_691 editions
1_323_401 documents 931_902 link sets
```

It seems reasonable that there might be more hundreds of thousands of
documents and millions of links in production. We can't find an obvious
fault in the existing subquery. Instead of risking a change to it we
create a new endpoint which does just the thing we need.

On the same data set we see a marked improvement:
```
2024-08-16 11:19:27 Document: 230366
2024-08-16 11:19:27 Edition: 230365
2024-08-16 11:19:27 Link: 230295
2024-08-16 11:19:27   0.000593   0.000170   0.000763 (  0.001478)
```

A negative side effect of this is that we won't get pagination and
sorting for free.

[1] https://github.com/alphagov/publishing-api/blob/3f4cc8523b1713aefc1e8f70950e1a81cfa56f22/app/presenters/queries/content_item_presenter.rb#L190