Fix Rails upgrade
We're seeing a degraded pod in integration with the following error:

```
Puma::HttpParserError: Invalid HTTP format, parsing fails. Are you trying to open an SSL connection to a non-SSL Puma?
```

For the upgrade to Rails on Publishing API, we accepted the change to
force SSL by default. We didn't do this on other apps we own. Given the
error mentions SSL, it's highly plausible that this is the cause

This updates the environment config files to match more closely the
result of Rails 7.2.0 upgrades in our other apps