PROOF OF CONCEPT: relaxing of specialist document schema validation
At the moment it is incredibly tedious to maintain the vast amount of document types and their facets for specialist documents. Unsure there is a lot of value in having these strict schema validations when a more relaxed one that doesn't check "valid values" for each field should be enough. This proof of concept works for publishing a specialist document.

NOTE: for the sake of spike, the schemas have been directly manipulated to remove the enum for "document_type". if build schemas is rerun, we will lose this configuration.
NOTE 2: another option is to remove enums entirely for document type across all schemas - this will also save us from having to add it to content_schemas/allowed_document_types.yml and content_schemas/formats/specialist_document.jsonnet . The question will then be whether or not that swings too far the other way.