Creating embed links will now look in any field
We’d like to support embeds in fields other than ‘body’.

In the previous iteration ‘body’ could only be a string or an array. Now it can be any field, there are more possible value types we need to support: hash, nested hash[2], bool[3], etc.

We now pass all the document details through to the find embed service. It’ll go through all fields to one level of nesting. More tree traversal may be needed but I couldn’t find any schema examples of this, nor did the test suite look to support them.

If the value isn’t an Array, Hash or String we ignore it in order to fail safely.

[2] https://github.com/alphagov/publishing-api/blob/da0c23f2b2abb83cfd1586a174858f2fc42670a0/content_schemas/formats/government.jsonnet#L19
[3] https://github.com/alphagov/publishing-api/blob/da0c23f2b2abb83cfd1586a174858f2fc42670a0/content_schemas/dist/formats/government/publisher_v2/schema.json#L178