Embed presenter works with multiple fields
Before if we had different fields with different content embedded, we
got an error because we were looping through all the linked editions,
and if a linked edition wasn’t present in a given field, we got an
error.

I’ve refactored the presented to fetch the embedded editions once, and
then loop through the actual content references when doing the find and
replace, which should work more reliably, as well as cutting down on
the amount of work the database needs to do.