Add licensify as a rendering app
This will allow us to publish a special route the ensures that we have
routing information for licensify in the Content Store. This supports
work to allow router to use the content store as source of truth for
routing information. Also in ensures that in case of disaster recovery,
we can easily republish routes to licensify.