Remove facet value enums from jsonnet
In an effort to reduce the number of code changes we need to make when configuring or editing a specialist finder, we're removing the rather verbose enums from the schemas.
We consider that validating that the facet values are strings is valuable enough, when measured against the cost of writing and maintaining the code.