Find embedded content by friendly_id
rather than content_id.

Now that we're storing an `EmbeddedContentReference` in Publishing API's
database with a `friendly_id`, we can allow users to enter this and look
up existing `EmbeddedContentReference`s by this value.

There's some not-so-nice naming going on in this file now that we have a
`EmbeddedContentReference` model and `EmbeddedContent` data object. This
will be cleaned up separately when we update the code to present
embedded content.