add a `content_block` update type to publish command
when a Content Block is updated, we want to support features that
relate to the updating of the Block and its Host Documents (Documents
that have the Block in their content, currently identified as Links).

As Content Blocks are a new type of content, we think it makes sense
that a new update type is issued, because we won't necessarily want to
follow the existing update flow for Editions and it semantically is a diffent
type of update - `major`/`minor` updates do not mean anything (currently) in
Content Block land.

This commit introduces the new update type without changing any of the
functionality related to `major` updates.