add a custom job to update Host Content
Now that we have an update type for Content
Blocks, we can call a custom job to update
dependencies when a Content Block is updated.

We only call this for Live Documents because for
Draft we currently don't need any special
behaviour.

For now we keep the same behaviour as the existing
DependencyResolutionJob, but trigger another new
update type of `host_content` - this will then
allow the email notification service to listen to
this update type and trigger emails for Host
Documents.

This is the same behaviour as the parent class,
except for the event type, so if we keep this
behaviour we could create a shared method.

I see two more changes that could be made in this
Job:
1. Only getting dependent Documents that embed the
relevant Block (a query similar to the one that
gets Editions in `GetEmbeddedContent`) so that we
only update Host Documents and not other types of
Links, e.g. Organisation.
2. We want to add Change Notes to Documents that
embed Blocks - we could do this here in this Job
and create a new ChangeNote for the edition on the
Publishing API database, and then add behaviour in
relevant publishers to sync their Change Notes back
up with Publishing.