Rename filename "*Worker" to "*Job"
It looks like this was missed when we renamed a bunch of files to
reflect the latest Sidekiq terminology in 88ea758312. The classname was
updated but the filename not, meaning the Job couldn't be found when
running the `spec/commands/put_publish_intent_spec.rb`.

I'm not quite sure how this made it through CI though if it's failing
locally...