Add `content_id_alias` to `content_block` schemas
We've made the `content_id_alias` an optional field for Content
Block types, but made it forbidden for any other types in the top-level
`default_format.jsonnet`. We've not made it required just yet to ensure
the flow for adding a new Content Block still works. We may want to
consider whether or not it should be required on the API side in future.

When submitted, we'll set this as the `name` of a new `ContentIdAlias` to
allow users to enter a human-readable alias when embedding content in an
Edition.

We discovered that to make this forbidden, we need to add a custom
method to define an `OptionalProperty` on the `SchemaGenerator::Format`
class to read this from the top-level `default_format` file. The
"forbidden" value is then overwritten by the lower-level Content Block
jsonnet file when the `derived_properties` are merged in. Before adding
this, the generator ignored this field and didn't throw an error when
passing this for non-Content Block types.

We're now fairly confident the schema is being validated correctly.

Co-authored-by: Ynda <yndajas@gmail.com>
Co-authored-by: Stuart Harrison <pezholio@gmail.com>