Create a `ContentIdAlias` on `PUT Content`
A `ContentIdAlias` will be created when making a request to this
endpoint when the API consumer submits the
`[:content_id_alias]` parameter. Per the preceding schema
updates, it will only be possible to provide a content ID alias for the
supported document types (email and postal addresses).

Co-authored-by: Ynda <yndajas@gmail.com>

TEMP: Check outside details hash for alias