Remove allowed_document_types
Tentative commit. We do not know if enforcing a doc type in all the jsonnets is desired.
None of the jsonnets are missing a doc type, so I'd be inclined to require it, and raise
if it's missing. There are some generic schemas only used for testing, where any string
can work as a type.
This would remove a step from "places where we need to configure stuff" for finders.