Update docs re: updating content schemas
These docs didn't explain the steps involved in adding new fields to the
`default_format.jsonnet` schema, which we had to do when adding
`content_id_alias` as a new field outside of those already declared
there.

This hopefully clarifies things, in particular the need to add new
methods to the `SchemaGenerator::Format` class which was hidden away.

Co-authored-by: George Eaton <george@dxw.com>