Add custom logic for document_type field
Specialist documents have a complex way of defining schemas, since they use a "generic" specialist_document schema, but also provide an enum of possible document types which needs to be maintained.

There is not enough value provided by the document type list, with enough maintenance cost to warrant relaxing the schema.

We have added a custom check in the generators, so that for specialist documents the generated schema only checks that the document type provided is a string. There is now no need to maintain the document type list, nor whitelisting the document type in the 'allowed_document_types' list.