Restore specialist doc types in allowed list
We've recently removed the specialist document types from the allowed list,
in an effort to relax the schema, so that it only checks that the document
type is a string.

There are places that use this list, and it should include
all the allowed types. We can keep the rest of our changes towards
relaxing the schema.