Add GraphQL world index types and query
This will allow us to get the data the frontend Collections app needs to
render the gov.uk/world page using the GraphQL endpoint

The `EditionType` here is very minimal, to make this change as small as
possible. We'll be adding additional fields to this in upcoming
branches, but these extra fields aren't really necessary for the
majority of the `/world` page in of itself, and there's a fair bit of
complexity involved in some of the Edition fields e.g. Links that needs
further thought.

The code in Types::QueryType::EditionTypeOrSubtype is a little complex:
- This class defines two types (with more to follow as we add more
  Edition subtypes) that can be used by the edition query. The
  `.resolve_type` method is used to distinguish between them using the
  `.base_path` method defined in Edition subtypes. The fields defined on
  the matching type will then be exposed.
- The way the base path argument is retrieved feels a little brittle,
  but it does the job for now. [A Stack Overflow answer](https://stackoverflow.com/a/62475738/4002016)
  was useful in working this out. The safe navigation operator before
  looking at the value was needed in order for consuming apps to read
  ('introspect'?) the schema, which gives them access to field/method
  definitions.

Co-authored-by: George <george@dxw.com>