Pass embeds into govspeak_attributes
Now the Govspeak gem supports embeds out of the box, we can remove the
ContentEmbedPresenter and pass the embedded_editions directly to
Govspeak.

As there’s a lot more going on in the embeds department in the
DetailsPresenter, I’ve created a different test file to (hopefully)
keep things neater and easy to follow.