WIP add additional edition/world index types
This expands the edition type to include all the top-level properties
that the `/world` content item contains, and most of the nested data.
The queries sent by frontend apps will be able to pare down the data
that gets sent to them from the Publishing API, but we are able to get
almost all the same data as presented by the current content item if
required

`links` needs further work to present the data that is currently
presented in the content item. `publishing_scheduled_at` and
`scheduled_publishing_delay_seconds` are not currently available in
Publishing API, so they just return `null`. This is what they currently
return for `/world` anyway, but for other pages, we might need to
explore how to retrieve these data if needed

This also moves non-edition-inheriting types inside the edition classes
that need them. This is to avoid a naming clash should be need to define
edition types with the same name later down the line, e.g. there exist
both world location editions and world location objects on the world
index edition, and these need different type definitions

Note that with the world index type, we have also provided top-level
access to the world locations and international delegations, while still
retaining a stringified copy of the `details` object, which is where
they were previously presented, and from where we build those top-level
fields

Co-authored-by: Ynda Jas <yndajas@gmail.com>