Move factory to live edition factory file
This looks like it belongs in the live edition factory, not the edition
factory. The fact that live editions have their own factory is a little
unusual and probably led to this misplacement

I've also changed it to be a trait rather than a specific factory. I'm
not sure that the distinction in how they're used is clear, but
`with_***` sounds trait-like