Use document_type when resolving subtypes
Here we use `document_type` rather than `base_path` when determining
whether to use a generic EditionType or one of its subtypes (in this
case, the WorldIndexType).

The need for this came about when adding Role GraphQL types for which
the `base_path` isn't a good identifier. `document_type` can refer to
multiple documents, meaning there's not a single `base_path` linked to
the type. For example, the `document_type` of `ministerial_role` can
refer to both the prime minister and other cabinet ministers, while the
base_path of `/government/ministers/prime-minister`, can only refer to
one role.

This also handily simplifies the code as we don't need to mess around in
GraphQL internals for the base path argument.

Co-authored-by: Ynda <yndajas@gmail.com>