Introduce RoleType for GraphQL prime minister page
We're planning to serve the prime minister page (from the Collections
frontend rendering app) by way of Publishing API's `/graphql` endpoint.

After some back-and-forth in the team, we've settled for defining only
the fields that're needed right now, for this one page. We can add more
(and extract the privately-declared types) as and when other queries
need it.

We've intentionally avoided optimising the database queries used here
(e.g. grouping them together in any kind of preloader). We assume that
that work will be driven by performance testing.

Note: we're not 100% convinced by the use of some of the existing
presenters in this new type. Like, maybe invoking the EditionPresenter
to process a single string is overkill. But, it appears to work and
saves us from duplicating logic or from causing too much disruption for
this still-experiment feature. Just noting.

Co-authored-by: Mike <mike.patrick2@digital.cabinet-office.gov.uk>
Co-authored-by: Ynda <yndajas@gmail.com>