Abstract presented edition retrieval in GraphQL
We've found a need to use this to fetch different pieces of data across
multiple types. Given the `Presenters::EditionPresenter#present` method
does quite a lot of work, it might be worth looking into reducing the
reliance on this, perhaps by moving some of its helper methods into some
kind of module. For now though, we're providing a quick, abstracted way
to get the presented edition on EditionType and its subtypes, and
memoising the return value in case we need use it to retrieve data for
multiple fields

Co-authored-by: George <george@dxw.com>
Co-authored-by: Mike <mike.patrick2@digital.cabinet-office.gov.uk>