Add a job to update statistics caches for embeds
This fetches all the links that have an embed type. We then loop
through them (in batches of 50) and fetch their associated editions
(and base path). We then send the query to get the pageviews and update
or insert the pageviews for each document as appropriate.

We then update the schedule to run at 2:30am every day to ensure the
data is up to date.