Refactor Embedded Content code
Having the query results passed to the presenter directly in the Query
object made the tests tricky to parse, additionally, we were using
`Edition.where` to carry out our `host_editions` query, which adds
attributes to an `Edition` which would not otherwise be there.

This commit adds a `GetEmbeddedContentService` which fetches the results
from the Query object and passes it to the presenter. We also update the
Query object to make a raw SQL query and cast it as a `Data` object
which better represents the results we get from our query.