DISTINCT by `unique_session_id`
The data we got back was quite different from what was present in the
Content Data service. After discussing with the Data team, we’ve found
that DISTINCTing by the `unique_session_id` (a combination of the
pseudonymous ID of a user on a particular client and the session ID
generated by Google Analytics) is the right way to go.