Log GraphQL query errors
I found it useful to see this information in Publishing API's logs. When
making bad requests from a simple client (e.g. curl), the response
itself contains all of this already, so it was hard to ignore. But when
making the request from a frontend app, all I could see was a 404 on the
frontend :)

But maybe I've missed somewhere that this is already surfaced for free?