Put GraphQL class self.document_type method at top
We did this in the role type class. It's nice to be able to see the
document type(s) the class implements near the top of the file. You can
easily see it in this class because it's small, but for larger type
classes like the role type class, that wouldn't be the case if defined
alongside other methods. This moves the definition in this class for
consistency