Setup Ministers Index for GraphQL endpoint
These type definitions allows our /graphql endpoint to serve the data
that the Collections frontend app uses when rendering the Ministers
Index page.

We've made a few, errm, pragmatic decisions here that it might be useful
to try to explain.

* Defining specialised types
As far as our familiarity with GraphQL goes, reusability seems to be
pretty central to how most teams use it. In contrast to that, these
first few content types that we're defining just need to serve the data
and get out of our way to enable testing the basics.

Development of these types has taken a lot of manual testing and
back-and-forth, checking content schemas, frontend code, the Publishing
API database, etc. So, the flexibility to make a daft number of changes
to these definitions without compromising the existing (role/person/etc)
definitions has been very useful.

FYI, during this work we ran into an issue where GraphQL Ruby appeared
to treat Ruby namespaces in a surprising way: it wouldn't allow us to
define, say, a Person type within two different namespaces. That's why
the names of these latest types look so terrible.

In the near future I think we'll want to work on de-duplicating a lot of
these types and their supporting code.

* Jumping through hoops to populate person->role_appointments->roles
The frontend app wants to consume data in which every person has an
associated set of links, those links... let me try to show you:

```
[...]
  "person": {
    ...,
    "details": { ... },
    "links": {
      "role_appointments": [
        {
          ...,
          "details": { ... },
          "links": {
            "role": [
              {
                ...,
                "details": { ... },
                "links": {}
              }
            ]
          }
        }
      ]
    }
  }
[...]
```

So, person->links->role_appointments->links->role. The frontend mostly
just needs a person's current roles, but it depends on the
role_appointments having a details->current=boolean, too.

So far we've mostly avoided thinking much about the database queries
that we're making in these type definitions. That will come in a later
piece of work, driven by performance tests.

Due to a quirk in how we worked on this particular type definition,
quite early on we ended up with a single database query to fetch
a person's current roles and then when we realised that these roles
would need to be nested inside links and role_appointments, we were
reluctant to split it up into multple database queries and to increase
the cost of the N+1s. Instead, we made use of the field option
`method: :itself` to pass-through the person object until we could
perform our single query. We hope that it makes enough sense to work
with later on.