Update embedded content query to fetch instances
This adds an `instances` field, which returns the amount of times a link
is referenced in a document. Because we now will potentially return
multiple rows for each document, we’ve had to add a `group_by` clause.
I’ve tweaked the `TABLES` constant to allow us to use it in the SELECT
and `GROUP BY` parts of the query.