Update spec to ensure instances are returned
I’ve tweaked this spec to be more “end-to-end”-y, so we’re actually
creating and publishing an instance via the API. This allows me to test
the instances are returned correctly throughout the entire lifecycle of
the journey, as I wasn’t 100% confident the instances would change if
the number of references changed. Happily they do!