Change World Index spec to compare hashes, not JSON string
The world index spec if currently comparing JSON strings, which means
the test has to specify the keys in exactly the same order as they are
included in the JSON.

Additionally, rspec doesn't give a useful diff when this test fails when
comparing JSON strings, like it would with a hash.

Therefore changing this spec to compare hashes, like we are doing in the
other GraphQL integration tests.