Move all edition fields onto a single edition type
This removes the hardcoding and duplication amongst the varieties of
Edition we have currently implemented.

These changes will make it easier to implement new document types, as
the fields and links have already been defined.

Since all links are now returned as the Edition type, any field can be
selected from them, not just those that have been defined in the link
type. This makes queries much more flexible, as the frontend application
can request any fields it likes from the links, without any changes
being made in Publishing API.

The Minister Index Type is being retained, since that contains custom
optimisations specific to the links on that page. In future, it may be
possible to decouple that type from editions entirely, meaning we
wouldn't need to ever publish that page from Whitehall.

Note 1: this requires selection of all fields within `details` for any
edition. Previously this was only required when `details` was nested
within a link, making it consistent whether the edition is returned at
the top-level or as a link.

Note 2: role appointment reverse links will be added in a separate
commit, as they require some logic.