Filter host content by content ID
When viewing the details of a particular piece of content that embeds
a block, we need to return the relevant metadata. Rather than returning
all the objects (which could be on multiple pages) and filtering through
them, we add a new endpoint that allows us to return a single content
item.