Fix rake task to unpublish Whitehall Frontend documents
This task did not work and returned an error when run in production.

This is because:

1. No locale was specified in the unpublish command, meaning the default
locale (English) was always attemted to be unpublished. In some cases,
it was not the English document we needed to unpublish, so this gave the
error `Could not find an edition to unpublish`.

2. Draft editions were not being removed before attempting to unpublish.
This gave the error `Cannot unpublish with a draft present`.

This updates the rake task to only remove the locale that actually needs
to be unpublished and also to discard the draft edition before
attempting to unpublish.