Use GraphQL-Ruby Dataloading for person's roles
This reduces the number of database queries made in serving the
Ministers Index page. With my local data the number went from 250 down
to 28.

The links between a person and their roles are indirect (i.e. not
a straightforward link or reverse link). If there are links on other
document types of this shape, it could make sense to extract the
ActiveRecord query into its own query definition under app/queries.
Right now this is quite specific to a person's current roles, though.