Return success from publish command when document already has a published edition and no draft exists
Previously the publish command raised a conflict error in this case.

However, this error was rarely useful to the Publishing apps, as in almost all cases it is caused by publishing apps having to do their own dependency resolution due to shortcomings in Publishing API's dependency resolution, or by Sidekiq's "at least once" job running guarantees.

We therefore return a success response from the publish command and perform a no-op in this situation, rather than return an HTTP 409 error response for the publishing app to handle.