Support multiple content types in details fields
According to ADR-004, it is possible to represent content in an array
format containing multiple types. This can include already parsed
govspeak in HTML format or only govspeak. In the latter case, Publishing
API parses this and converts to HTML before sending the content to
Content Store.

This replicates that behaviour for GraphQL responses, expanding this to
all fields within `details`, not just the body (e.g. the
`more_information` and `need_to_know` fields, which use the Content
Schema defition `body_html_and_govspeak`)..

It also has the side effect of supporting embedded content and full
change history in GraphQL responses.