Add index for edition links
We are currently querying on `edition_id` and `link_type` in GraphQL's
`LinkedToEditionSource` dataloader.

At the moment, we are only using an index for the `edition_id` part of
the query, not the `link_type` part as well.

Adding a composite index to allow one index to handle both parts of this
query.