Make migrations concurrent
The migrations to add new indexes were previously locking all writes to
the editions table at the time of running.

As one of the migrations is slow (several minutes), this caused degraded
performance of the application.

Making them concurrent to remove the need for the tables to be locked.

Concurrent indexes must be created outside of a transaction [1].

1: https://thoughtbot.com/blog/how-to-create-postgres-indexes-concurrently-in