Only retrieve change history from database if needed
At the moment, we are retrieving the full change history for all
documents from the datbase, whether it is needed or not.

This makes a change so the change history is only retrieved from the
database if it has actually been requested by the client.

For the prime minister page, the number of database queries is reduced
from 254 to 93.