pass embed code to Content Block presenter
with the latest update to the ContentBlockTools
gem we are enabling field names to be passed in the
embed code. So e.g.

{{embed:content_block_postal_address:2593f06c-b7c8-4173-9a47-09d9bdd767ef/town_or_city}}

would return just the `town_or_city` field on a postal
address content block. Field names currently are
restricted to those within the `details` json blob on a
block.